+++ packages/transaction-controller/src/helpers/PendingTransactionTracker.test.ts
@@ -5,6 +5,7 @@ import { freeze } from 'immer';
 
 import { PendingTransactionTracker } from './PendingTransactionTracker';
 import { TransactionPoller } from './TransactionPoller';
+import type { TransactionControllerMessenger } from '../TransactionController';
 import type { TransactionMeta } from '../types';
 import { TransactionStatus } from '../types';
 
@@ -74,11 +75,25 @@ function createTransactionPollerMock(): jest.Mocked<TransactionPoller> {
   } as unknown as jest.Mocked<TransactionPoller>;
 }
 
+/**
+ * Creates a mock messenger instance.
+ *
+ * @returns The mock messenger instance.
+ */
+function createMessengerMock(): jest.Mocked<TransactionControllerMessenger> {
+  return {
+    call: jest.fn().mockReturnValue({
+      remoteFeatureFlags: {},
+    }),
+  } as unknown as jest.Mocked<TransactionControllerMessenger>;
+}
+
 describe('PendingTransactionTracker', () => {
   const queryMock = jest.mocked(query);
   let blockTracker: jest.Mocked<BlockTracker>;
   let pendingTransactionTracker: PendingTransactionTracker;
   let transactionPoller: jest.Mocked<TransactionPoller>;
+  let messenger: jest.Mocked<TransactionControllerMessenger>;
 
   let options: jest.Mocked<
     ConstructorParameters<typeof PendingTransactionTracker>[0]
@@ -112,6 +127,7 @@ describe('PendingTransactionTracker', () => {
   beforeEach(() => {
     blockTracker = createBlockTrackerMock();
     transactionPoller = createTransactionPollerMock();
+    messenger = createMessengerMock();
 
     jest.mocked(TransactionPoller).mockImplementation(() => transactionPoller);
 
@@ -123,6 +139,7 @@ describe('PendingTransactionTracker', () => {
       getTransactions: jest.fn(),
       getGlobalLock: jest.fn(() => Promise.resolve(jest.fn())),
       publishTransaction: jest.fn(),
+      messenger,
     };
   });